add food item deprecation function
