+++ src/blocks/food-and-drinks/food-item/deprecated/deprecatedIconsToBackgroundImage.js
@@ -1,7 +1,15 @@
 /**
- * WordPress dependencies
+ * Internal dependencies
  */
-import { SVG, Path } from '@wordpress/components';
+import { hasEmptyAttributes } from '../../../../utils/block-helpers';
+import fromEntries from '../../../../js/coblocks-fromEntries';
+
+/**
+ * WordPress dependencies.
+ */
+import { __ } from '@wordpress/i18n';
+import { RichText } from '@wordpress/block-editor';
+import { Icon, SVG, Path } from '@wordpress/components';
 
 /**
  * Block user interface icons
@@ -10,9 +18,7 @@ const icons = {};
 
 icons.glutenFree = (
 	<SVG
-		height="14"
 		viewBox="0 0 14 14"
-		width="14"
 		xmlns="http://www.w3.org/2000/svg"
 	>
 		<Path d="m6.99766252.00002487c-3.85852931 0-6.99766252 3.13913321-6.99766252 6.99766252s3.13913321 6.99766251 6.99766252 6.99766251c3.85857908 0 6.99771228-3.1391332 6.99771228-6.99766251s-3.1391332-6.99766252-6.99771228-6.99766252zm-.16527708 8.72190053c-.1740675.18691119-.36012079.35314565-.65887034.49870337-.29879929.14560746-.67575489.21838011-1.13207283.21838011-.58016696 0-1.06856305-.18033392-1.43575843-.54106394-.36724512-.36068029-.56548313-.82895915-.56548313-1.40479929v-.99086679c0-.5736643.19144938-1.04139609.54459503-1.40318295.35308348-.36179928.81523268-.54269271 1.37153996-.54269271.57799112 0 1.02444938.14072113 1.33193961.42208881.30746536.28142984.46633925.6577016.47285435 1.11614565l-.00465009.02587389h-.89634813c-.01741918-.27353464-.0977762-.45774778-.24119538-.60442452-.14340675-.14666429-.35313322-.22316696-.62905507-.22316696-.29121492 0-.51775133.1107691-.69588454.33546537-.17819538.2247087-.25912434.5123428-.25912434.86619715v.99646181c0 .36036945.08417407.65290231.26889698.87761102.18467318.22469627.42943695.33699467.74233571.33699467.22163765 0 .40621136-.02271581.54097691-.06824689.13470337-.04548135.20091119-.10180462.31281172-.16898224v-.86832327h-.88277087v-.6589698h1.81527531v1.78079751zm4.11545296-3.37229663h-2.36234462v1.31793961h2.02664302v.70870338h-2.02664302v1.97690941h-.93250444v-4.71225577h3.29484906z" />
@@ -21,9 +27,7 @@ icons.glutenFree = (
 
 icons.vegetarian = (
 	<SVG
-		height="18"
 		viewBox="0 0 18 18"
-		width="18"
 		xmlns="http://www.w3.org/2000/svg"
 	>
 		<Path
@@ -35,9 +39,7 @@ icons.vegetarian = (
 
 icons.spicy = (
 	<SVG
-		height="18"
 		viewBox="0 0 18 18"
-		width="18"
 		xmlns="http://www.w3.org/2000/svg"
 	>
 		<Path
@@ -50,9 +52,7 @@ icons.spicy = (
 
 icons.pescatarian = (
 	<SVG
-		height="18"
 		viewBox="0 0 18 18"
-		width="18"
 		xmlns="http://www.w3.org/2000/svg"
 	>
 		<Path d="m17.8325581 4.90148962c-.1046511.04822471-2.344186.91626951-3.9767441 2.12188728-.711628-.94038186-1.7162791-1.88076372-2.9302326-2.55590967-.8162791-1.47085369.3348837-2.79703324.3348837-2.79703324-2.34418603-.6992583-4.10232557.89215716-5.0232558 2.00132551-3.45348837.86804479-6.09069767 4.89480815-6.09069767 5.69051588 0 .62692124 3.07674418 5.47350472 6.96976744 5.81107762.52325581.675146 1.59069767 1.6396402 2.99302323 1.2297302 0 0-.66976742-.747483-.23023253-1.5914155 1.75813953-.6028089 3.20232553-1.7843143 4.10232553-3.0381568 1.611628 1.1573931 3.7465117 2.0013255 3.8720931 2.0495502-.0418605-.1446741-.9627907-3.327505-.9627907-4.46078572-.0209303-1.66375252.9209302-4.36433633.9418604-4.46078576zm-13.8359579 4.14672815c-.54545455 0-1-.45454545-1-1 0-.54545454.45454545-1 1-1s1 .45454546 1 1c0 .56363637-.45454545 1-1 1z" />
@@ -67,9 +67,7 @@ icons.popular = (
 
 icons.vegan = (
 	<SVG
-		height="18"
 		viewBox="0 0 18 18"
-		width="18"
 		xmlns="http://www.w3.org/2000/svg"
 	>
 		<Path
@@ -80,4 +78,136 @@ icons.vegan = (
 	</SVG>
 );
 
-export { icons as oldIcons };
+const isEmpty = attributes => {
+	const attributesToCheck = [ 'url', 'title', 'description', 'price' ];
+	const newAttributes = Object.entries( attributes ).filter( ( [ key ] ) =>
+		attributesToCheck.includes( key )
+	);
+
+	if ( ! Object.fromEntries ) {
+		return hasEmptyAttributes( fromEntries( newAttributes ) );
+	}
+
+	return hasEmptyAttributes( Object.fromEntries( newAttributes ) );
+};
+
+export default function deprecatedIconsToBackgroundImage( { attributes } ) {
+	return isEmpty( attributes ) ? null : (
+		<div
+			className={ attributes.className }
+			itemScope
+			itemType="http://schema.org/MenuItem"
+		>
+			{ !! attributes.showImage && attributes.url && (
+				<figure className="wp-block-coblocks-food-item__figure">
+					<img
+						src={ attributes.url }
+						alt={ attributes.alt }
+						itemProp="image"
+						style={ {
+							objectPosition: attributes.focalPoint ?
+								`${ attributes.focalPoint.x * 100 }% ${ attributes.focalPoint.y *
+										100 }%` :
+								undefined,
+						} }
+					/>
+				</figure>
+			) }
+			<div className="wp-block-coblocks-food-item__content">
+				<div className="wp-block-coblocks-food-item__heading-wrapper">
+					<RichText.Content
+						tagName="h4"
+						className="wp-block-coblocks-food-item__heading"
+						value={ attributes.title }
+						itemprop="name"
+					/>
+					{ ( !! attributes.spicy ||
+						!! attributes.vegetarian ||
+						!! attributes.glutenFree ||
+						!! attributes.pescatarian ||
+						!! attributes.popular ||
+						!! attributes.vegan ) && (
+						<div className="wp-block-coblocks-food-item__attributes">
+							{ !! attributes.popular && (
+								<span className="hint--top" aria-label={ __( 'Popular', 'coblocks' ) }>
+									<Icon
+										icon={ icons.popular }
+										className="wp-block-coblocks-food-item__attribute wp-block-coblocks-food-item__attribute--popular"
+									/>
+								</span>
+							) }
+							{ !! attributes.spicy && (
+								<span className="hint--top" aria-label={ __( 'Spicy', 'coblocks' ) }>
+									<Icon
+										icon={ icons.spicy }
+										className="wp-block-coblocks-food-item__attribute wp-block-coblocks-food-item__attribute--spicy"
+									/>
+								</span>
+							) }
+							{ !! attributes.spicier && !! attributes.spicy && (
+								<span className="hint--top" aria-label={ __( 'Spicier', 'coblocks' ) }>
+									<Icon
+										icon={ icons.spicy }
+										className="wp-block-coblocks-food-item__attribute wp-block-coblocks-food-item__attribute--spicier"
+									/>
+								</span>
+							) }
+							{ !! attributes.vegetarian && (
+								<span className="hint--top" aria-label={ __( 'Vegetarian', 'coblocks' ) }>
+									<Icon
+										icon={ icons.vegetarian }
+										className="wp-block-coblocks-food-item__attribute wp-block-coblocks-food-item__attribute--veg"
+									/>
+								</span>
+							) }
+							{ !! attributes.glutenFree && (
+								<span className="hint--top" aria-label={ __( 'Gluten Free', 'coblocks' ) }>
+									<Icon
+										icon={ icons.glutenFree }
+										className="wp-block-coblocks-food-item__attribute wp-block-coblocks-food-item__attribute--gf"
+									/>
+								</span>
+							) }
+							{ !! attributes.pescatarian && (
+								<span className="hint--top" aria-label={ __( 'Pescatarian', 'coblocks' ) }>
+									<Icon
+										icon={ icons.pescatarian }
+										className="wp-block-coblocks-food-item__attribute wp-block-coblocks-food-item__attribute--pescatarian"
+									/>
+								</span>
+							) }
+							{ !! attributes.vegan && (
+								<span className="hint--top" aria-label={ __( 'Vegan', 'coblocks' ) }>
+									<Icon
+										icon={ icons.vegan }
+										className="wp-block-coblocks-food-item__attribute wp-block-coblocks-food-item__attribute--vegan"
+									/>
+								</span>
+							) }
+						</div>
+					) }
+				</div>
+				<RichText.Content
+					tagName="p"
+					className="wp-block-coblocks-food-item__description"
+					value={ attributes.description }
+					itemprop="description"
+				/>
+				{ !! attributes.showPrice && attributes.price && (
+					<p
+						className="wp-block-coblocks-food-item__price"
+						itemProp="offers"
+						itemScope
+						itemType="http://schema.org/Offer"
+					>
+						<RichText.Content
+							tagName="span"
+							value={ attributes.price }
+							itemprop="price"
+						/>
+					</p>
+				) }
+			</div>
+		</div>
+	);
+}
